Moved Request extensions for Response Serialization into a separate Swift file.
